One day, going through old books in the attic, schoolboy Ivan found an English-Latin dictionary. He knew English perfectly by then, and his dream was to learn Latin. So the dictionary he found was just in time.
Unfortunately, for a full-fledged study of the language only one dictionary was not enough: in addition to the English-Latin dictionary, he needed the Latin-English one. For lack of a better one, he decided to make a second dictionary out of the first one.

As you know, a dictionary consists of translatable words, each of which has several translations. For each Latin word found anywhere in the dictionary, Ivan proposes to find all of its translations (that is, all English words for which our Latin was found in his list of translations), and consider them and only their translations of this Latin word.

Help Ivan do the work of creating a Latin-English dictionary from the English-Latin dictionary.

The first line contains a single integer N - the number of English words in the dictionary. This is followed by N descriptions. Each description is on a single line, first the English word, then a space-separated hyphen (symbol number 45), then comma-separated translations of the English word into Latin. The translations are sorted in lexicographical order. The order of the English words in the dictionary is also lexicographic.

All words consist only of small Latin letters and the length of each word does not exceed 15 characters. The total number of words in the input does not exceed 100000.

On the first line print the number of words in the appropriate Latin-English dictionary. From the second line print the dictionary itself, keeping exactly the format of the input data. Specifically, the first line should be the translation of the lexicographically minimal Latin word, then the second in this order, and so on. Within the translation, the English words should also be sorted lexicographically.