The teacher gave Petya a homework assignment - in a given text to put the accents in the words, and then asked Vasya to check this homework. Vasya is very unfamiliar with the topic, so he found a dictionary, which indicates how to place the accents in words. Unfortunately, not all the words are in this dictionary. Vasya decided that in words that are not in the dictionary, he would consider that Petya put the accent correctly, if Petya put exactly one accent in this word.
It turned out that in some words the accent can be placed in more than one way. Vasya decided that in this case, if the way Petya placed the accent corresponds to one of the variants given in the dictionary, he will count it as a correct accentuation, and if it does not correspond, it will be counted as a mistake.

You are given the dictionary Vasya used and the homework Petya turned in. Your task is to find the number of mistakes that Vasya will count in this homework.

First enter the number N - the number of words in the dictionary (0 ≤ N ≤ 20000).

Then comes N lines with words from the dictionary. Each word consists of no more than 30 characters. All words consist of small and capital Latin letters. There is exactly one uppercase letter in each word - the one on which the accent is placed. Words in the dictionary are arranged in alphabetical order. If there are several possibilities for placing the accent in the same word, these variants in the dictionary are in random order.

Next is an exercise done by Petya. The exercise is a line of text, not more than 300000 characters in total. The line consists of words that are separated from each other by exactly one space. The length of each word does not exceed 30 characters. All words consist of small and capital Latin letters (capital letters are the letters above which Petya put the accent). By mistake Petya could have put more than one accent in some word or not put the accent at all.