The shoe store sells shoes in different sizes. It is known that one pair of shoes can be put on top of another if they are at least three sizes larger. A customer comes to the store. You have to determine the greatest number of pairs of shoes the seller can offer him so that he can wear them all at once.
First you enter the size of the buyer's feet (he can not wear shoes of smaller size), then the number of pairs of shoes in the store and the size of each pair. The size is a natural number which does not exceed 100, the number of pairs of shoes in the store does not exceed 1000.

Print a single number - the maximum number of pairs of shoes.