During the Olympiad each participant received his identification number - a natural number. You need to sort the list of participants of the Olympiad by the number of points they gained.
The first line contains a number N (1 ≤ N ≤ 1000) - the number of participants. On each next line the identification number and the accumulated number of points of the corresponding participant is given. All numbers in the input file do not exceed 105.

In the output file print the original list in descending order of scores. If some participants have the same points, then they must be arranged among themselves in ascending order of identification number.