Output all source points in order of their distances from the origin.
The input for the program is a set of points on the plane. First the number of points n is given, then comes a sequence of n lines, each containing two numbers: the coordinates of a point. The value of n doesn't exceed 100, all the initial coordinates are integers, not exceeding 103.

You should print all the initial points in ascending order of their distances from the origin.