Print the names of the students in order of their grade point average.
Given first the number of students n, then n lines, each containing a surname, a first name and three numbers (marks in three subjects: mathematics, physics, computer science). The data in a line is separated by one space. The grades take on a value from 1 to 5.

You need to output a pair of first and last name, one per line, separating the name and last name with a single space. You don't need to print the grades. If several students have the same grade point average, they should be printed in the order given in the input.